net-dns/maradns: fix totally broken patch
In commit 73937e86b92dcf813bf021a83451e2ae4dbc423c, a version bump was
applied, and a patch file in FILESDIR which was applied to the old
version was "rebased" against the new release and applied in-place for
both the old and new versions.

Result: stable ebuild does not compile because patch doesn't match.

Fixes: 73937e86b92dcf813bf021a83451e2ae4dbc423c
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>