app-i18n/xvnkb: error out immediately if no Makefile exists
This package is absolutely horrible. It has the world's worst homebrew
"configure" script, which fails but doesn't *error out* with
-Wimplicit-int as a compiler default error, and also has:

config.h:1:1: error: expected identifier or ‘(’ before ‘-’ token
    1 | -e #ifndef __VK_CONFIG_H

which is just... dandy... I too love unix command flags echo'ed into my
headers...

It's a wreck but we can at least detect when it cannot compile anything
and abort with a clearer message.

Bug: https://bugs.gentoo.org/900398
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>