net-dns/maradns: update filter-lto comment
Upstream applied a patch which should fix it. I haven't tested, since it
doesn't backport cleanly. The distfile randomly renames a directory --
the one with the patched code! -- from "deadwood-github" in git to
"deadwood-${PV}" in the distfile, and I cannot be bothered to put up
with such shenanigans and hand-edit a patch hoping to get it to apply.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>