dev-util/cucumber-html-formatter: update SRC_URI
Use the gem since it contains the precompiled npm assets. Unfortunately
the gem no longer contains the specs, so we are not able to run them
anymore. We could use the upstream git repository as a source but that
would require creating and distributing the assets ourselves for each
release.

Signed-off-by: Hans de Graaff <graaff@gentoo.org>