media-video/rav1e: unvendor libgit2 from rust dependency
It is actually only used at build time, by this... thing:
https://crates.io/crates/built

Apparently it is of the utmost importance that when you get the version
info from your git checkout of the project, you do so with blazing-fast
fearless concurrency, and if that means spending several minutes
compiling a git implementation instead of running /usr/bin/git, then no
price is too large.

Fearless concurrency for the win! Onward!

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>