dev-qt/qtwebengine: add IUSE=webdriver to 6.7+
No deps but make it optional given most people do not need this
and it adds a bit of build time plus ~35MB to the install.

For the IUSE name, debated IUSE=webenginedriver as well but "webdriver"
is the name of the specification, and it felt redundant to have
webengine in the IUSE name for the webengine package.

wrt tests, unfortunately like most tools-related tests (see qttools),
it tries to use the system's tool and makes it difficult to specify
a path -- so skipping tst_webenginedriver for simplicity, it's not
a component that need to be overly worried about either way (haven't
tried but imagine may be further messy with sandbox too).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>