sys-libs/libsemanage: Make python optional
This commit adds an enabled-by-default python USE flag that gates the
build of the python bindings of the libsemanage library. The commit
also moves the additions to the config file to a patch instead of
doing it inline in src_prepare.

Reasons for making python optional is that in Flatcar, which is based
on Gentoo, we don't have python interpreter in the production images,
so we had our own fork of libsemanage ebuilds where we were dropping
all the python-related bits [1]. With this commit, we are trying to
make our changes less ad-hoc.

And for the config additions - in Flatcar we want to enable module
compression, which is specified by the ebuild as disabled. Our idea
was to make the modification as a user patch [2], but for this to
work, the additional config needed to be already made by the time the
user patches are applied. Thus moving the config to a patch too.

[1]: https://github.com/flatcar/scripts/commit/90bd28e13ff4b40820ab330478acf463ab54c972

[2]: https://github.com/flatcar/scripts/commit/90bd28e13ff4b40820ab330478acf463ab54c972

Signed-off-by: Krzesimir Nowak <knowak@microsoft.com>