sys-apps/policycoreutils: Make python and Gentoo extras optional
This commit adds an enabled-by-default python USE flag and
disabled-by-default vanilla USE flag.

Vanilla USE flag is for deciding to use just upstream SELinux
userspaces sources or to additionally pull a tarball with extra
utilities from Gentoo servers. As the extra tarball contains at the
moment a utility written in Python, disabling vanilla builds requires
enabling python support.

I'm not exactly sure why this package is pulling in app-admin/setools,
so I assumed that the script from the extra tarball requires something
from app-admin/setools. Thus made the dependency optional too.

Reasons for adding those USE flags is that in Flatcar, which is based
on Gentoo, we don't have python interpreter in the production images,
so we had our own fork of policycoreutils ebuilds where we were
dropping all the python-related bits and the extra scripts [1]. With
this commit, we are trying to make our changes less ad-hoc.

[1]: https://github.com/flatcar/scripts/commit/a38d44bd1c119ddb60e16a0718ddcd26ace6f8b4

Signed-off-by: Krzesimir Nowak <knowak@microsoft.com>