media-sound/din: mark as LTO-unsafe
The upstream website has a couple quirks. The open source release
tarball is "Download the complete source code from Internet Archive",
and there's no bug tracker or revision control. If the author uses
either one, I suppose it is internal...

There *is* a clearly publicized contact email, as well as e.g. requests
of:

> If you can package DIN Is Noise for various GNU/Linux distributions or
> other Unices, please e-mail me.

So my assumption is that email is the best way to report such issues.
Apologies in advance for no one else being able to easily check up on
the status of the bug report. ;) I have duly sent a personal email to
the author with a bug report.

Closes: https://bugs.gentoo.org/860513
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>