sys-devel/binutils: pass --disable-gdbserver too
Note that this doesn't matter for non-live, because gdbserver isn't included
in the dist tarball.

I only noticed it with live before binutils-2.42 release, then forgot about
it as binutils-2.42 proper was fine. Anyway, it causes a build failure with
USE=pgo + 9999, but it's of course completely pointless for this ebuild anyway.

Signed-off-by: Sam James <sam@gentoo.org>