Revert "sys-devel/gcc: drop obsolete cross fixincludes patch for 13"
This reverts commit 456eee6e374b50a4d9108a642ca29c5573ac5cb8.

Restore the fixincludes revert. The situation has improved a bit though,
because we now only run fixincludes when it's truly needed, so having
this patch is less risky to begin with.

Caveat emptor for trying to remove this in future: make sure to test
clean builds without mingw already built, and with.

Bug: https://bugs.gentoo.org/905118
Bug: https://bugs.gentoo.org/925204
Closes: https://bugs.gentoo.org/926059
Signed-off-by: Sam James <sam@gentoo.org>