dev-libs/libsecp256k1: fix build when CC is overridden to clang
The libsecp256k1 build system detects compiler flag feature support
using ${CC} but then builds the build-time executables using
${CC_FOR_BUILD}. When CC and CC_FOR_BUILD refer to different brands or
versions of compilers, the feature detection for one will not be
applicable to the other. Previous to this commit, our cross-compilation
support patch attempted to override the flags when building the
build-time executables, but it was doing so in a way that Automake does
not respect. This commit fixes the patch so that Automake will actually
use CFLAGS_FOR_BUILD, CPPFLAGS_FOR_BUILD, and LDFLAGS_FOR_BUILD when
building the build-time executables.

Closes: https://bugs.gentoo.org/895048
Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>