dev-util/maturin: Build maturin executable via cargo.eclass
Split the build into building the `maturin` executable via cargo.eclass,
and the Python modules via `setuptools`, in order to avoid unnecessarily
rebuilding it multiple times.  In order to achieve this, we strip
`rust_extensions=` from `setup.py` — we can also remove the dependency
on `dev-python/setuptools-rust` while at it.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/35671
Signed-off-by: Michał Górny <mgorny@gentoo.org>