linux-info.eclass: improve binpkg handling
When installing a binpkg, Portage will restore the ebuild environment
from when the binpkg was built. The stored values for KV_* may be
different from the values that would be detected on the current system.

To work around this, unset some key variables in
linux-info_get_any_version when MERGE_TYPE == binary.

Also, only look for the running kernel version: we can't possibly need
a full kernel source tree when installing a package that has already
been compiled. This gives a nice speedup as well.

Closes: https://bugs.gentoo.org/926063
Signed-off-by: Mike Gilbert <floppym@gentoo.org>