x11-drivers/xf86-video-openchrome: mark LTO-unsafe, strict-aliasing unsafe
Freedesktop moved to gitlab and dropped support for previous project
hosting methods to read-only. It appears openchrome never made the
migration. May be dead -- it hasn't been developed in some time, the
cgit HEAD is 4 years old and a non-default branch is "only" a year old.

Not really sure where to report a bug to. ;)

Closes: https://bugs.gentoo.org/864406
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>