dev-db/slony1: fix typo in ebuild variables
REQUIRE_USE isn't a thing. Due to this failing to be set properly, the
build faild during setup with:

 * ERROR: dev-db/slony1-2.2.10-r1::gentoo failed (setup phase):
 *   One of the postgres_targets_postgresSL_OT use flags must be enabled

I'm not sure why that error message exists at all though, since it's
supposed to be handled before setup via REQUIRED_USE. And... it usually
would be. ;)

Make this sane again.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>