dev-util/buildbot*: reinstate ~riscv keywords dropped during update
dev-util/buildbot has a new dependency in 3.11 which is missing this
keyword. So dropping keywords here makes sense, and same with other
packages that depend on buildbot (master).

But not all buildbot* packages depend on buildbot (master), so there
wasn't really any reason to drop keywords there...

In particular, the worker has no particular reason to be excluded from
running on riscv and connecting to a master on amd64.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/35699
Signed-off-by: Sam James <sam@gentoo.org>