dev-libs/rocr-runtime: remove incorrect sed that doesn't apply
 * SED: the following did not cause any changes
 *     sed -i -e "s:1.7.0:${PV}:" CMakeLists.txt || die;
 * no-op: -e s:1.7.0:5.7.1:

I don't actually understand the reasoning for attempting to modify it at
all, since the cmake function that this is an argument to (before it
changed to 1.8.0) explicitly takes the value, splits it on the periods
and adds each component to a VERSION_MAJOR etc cmake variable.

Then it also runs git, regardless of the input value, and tries to
calculate VERSION_COMMIT_COUNT.

So, there was nothing to fix as far as I can tell here? The only thing
sed'ing this out would ever do AFAICT is change the soname of libraries
from "1.7.0" to whatever the current ${PV} is... which is NOT how
sonames work, and upstream had a good reason to NOT have this be ${PV}.

It's quite irrelevant now since clearly the package worked just fine
despite the sed not applying, thus it *cannot* be fixing anything, and
also thus, there is clearly nothing to fix.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>