dev-qt/qtwebengine: skip tests newly failing with offscreen in 6.7+
They attempt to open display regardless and get a SIGTRAP ending
the entire test. If there is a display but it is still =offscreen,
the tests continue but with some normal failures instead. All good
with =xcb and a display.

Skip for now, albeit may consider virtx if it stays like that
(or if more tests start failing) given it's rather major tests.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>