app-text/texlive-core: filter LTO, no-SA for 2021
Not doing 2023 as not tested it yet. 2021 at least has a type mismatch
failure in some pascal to C thing (p2c) and there's also aliasing violations
in utils/pmx.

We should report these upstream as Karl is very nice & accommodating but no
point in doing so with an old TL.

Closes: https://bugs.gentoo.org/915223
Signed-off-by: Sam James <sam@gentoo.org>