sys-libs/libomp: Block calling nvptx-arch
Block calling nvptx-arch as well as amdgpu-arch.  Upstream is using
this executable to determine the GPU architecture and implement
the "native" card selection.  However, it is called even if the option
is not used (we are following upstream in building support for all GPUs)
and it may cause sandbox violations.

Closes: https://bugs.gentoo.org/926613
Signed-off-by: Michał Górny <mgorny@gentoo.org>