dev-libs/opensc: fix test search path for softhsm, disable broken
These tests are currently getting run only when dev-libs/softhsm is
installed, BUT on some architectures (including amd64), they use the
wrong search path (we install in /usr/$(get_libdir)/softhsm) and thus
think it's not installed anyway and skip.

Firstly, this commit enables these tests by overriding the search path.
However, one of the tests is broken (all architectures).  This seems to
be at least partially known upstream as there is a long trail of commits
disabling it for each individual release tag; but upstream seems to
recently have stopped disabling it for unknown reasons.  Given the
flakiness, disable this specific test downstream.

Bug: https://bugs.gentoo.org/909781
Signed-off-by: Matoro Mahri <matoro_gentoo@matoro.tk>