dev-qt/qtwayland: do qtbase:6[accessibility=] matching now
Backported from live ebuilds, albeit still skipping it for qtwebengine
given few users should be affected and there should be the 6.6.3
release in bit over a week (not worth a rebuild now).

Mostly a side effect of kwin-6 (formerly) requiring it so users
enabled it only for it and then disable after it no longer is.
Typical accessibility users likely leave it enabled globally.

Closes: https://bugs.gentoo.org/926992
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>