eclass/selinux-policy-utils.eclass: New eclass
The eclass is an attempt to eventually get rid of code duplication
between selinux-policy-2.eclass and sec-policy/selinux-base-policy
ebuilds. The code tries to avoid using any global variables, thus
making fewer assumptions about e.g. locations of policy files or
patches at the expense of passing more parameters to functions.

The commit also tries to use bash arrays in more places, so module
lists, policy types, etc. are taken as bash arrays.

Third change is a fix of handling the output from qdepends that we use
to get reverse dependencies of the merged SELinux policy
package. qdepends prints a name of the dependency and an RDEPENDS
snippet showing how the package pulled in the policy
package. Filtering out lines that contain the name of the policy
package from such output always results in an empty output. Process
the qdepends output a bit to receive the correct results.

Fourth change the commit makes is replacing the use of rlpkg script
with tools that we already depend on. So far we were using only a part
of rlpkg functionality, namely listing the contents of the passed
packages and then reseting the security context of the contents using
restorecon. This could be replaced using qlist or equery. This has an
advantage of not requiring python in SELinux profile as rlpkg was
written in python. This is not done in this commit though.

Signed-off-by: Krzesimir Nowak <knowak@microsoft.com>