dev-libs/folks: add 0.15.7, fix tests handling
Intentionally not yet 0.15.8 as upstream mirrored tarball was accidentally
made bad and equals 0.15.7 tag instead of the 0.15.8 tag.
Fix tests passing when they shouldn't (missing die), avoid bogus warnings
in tests when flatpak is installed and fix tests (that now would fail due
to the fix to make them fail when they don't pass) with
python-dbusmock-0.30.1 or newer.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>