media-libs/libopenraw: remove typo
The libtool flag is '-all-static' and not '--all-static' which causes
slibtool to fail while GNU libtool silently ignores the unknown flag.
Additionally fixing the typo causes the build to fail with undefined
references so just remove it instead.

Upstream already removed it themselves, but their change is not trivial
to backport.

Closes: https://bugs.gentoo.org/913723
Upstream-Commit: https://gitlab.freedesktop.org/libopenraw/libopenraw/-/commit/eb873c75b09074d710e3f105b9c988a28058db53
Signed-off-by: orbea <orbea@riseup.net>
Closes: https://github.com/gentoo/gentoo/pull/35789
Signed-off-by: Sam James <sam@gentoo.org>