dev-lang/rust: Disable LTO by default and use thin
We've had a few issues with lto on, and it isn't on by default generally
among Gentoo packages, so this commit flips it to be off by default.

Additionally, in an upstream ticket[0] it was suggested to use thin LTO
rather than fat LTO, so this commit makes that adjustment as well.

[0] https://github.com/rust-lang/rust/issues/121124

[sam: Note that there's a risk of miscompilations with non-thin LTO, per the
upstream bug(s).]

Bug: https://bugs.gentoo.org/924301
Signed-off-by: Randy Barlow <randy@electronsweatshop.com>
Closes: https://github.com/gentoo/gentoo/pull/35796
Signed-off-by: Sam James <sam@gentoo.org>