media-libs/zmusic: mark as LTO-unsafe
It's not actually LTO unsafe. Its bundled adlmidi dependency is.
Unfortunately, this package uses bundled dependencies and it's only
gotten worse in later upstream releases.

Issue still present in upstream git and in the libadlmidi source as
well.

Not currently up to figuring out how to patch the heck out of this to
devendor it, given upstream doesn't provide options for it.

Bug: https://github.com/ZDoom/ZMusic/issues/56
Closes: https://bugs.gentoo.org/860117
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>