gstreamer-meson.eclass: Avoid needlessly compiling helper libraries in split packages
When a split package uses gstreamer_system_package() for a helper library in gst-libs/
that is in the same tarball, it wasn't working fully anymore these days because for
subprojects support they all do a meson.override_dependency as well, which is
also overriding the dependency our eclass helper adds at the top of the split plugins
subdir meson.build. Because of that, it was building these helper libraries from
scratch and link against them, only to discard it after linking against it and
potentially also causing some ABI issues, as at runtime the system version from the
main package is used (though that hasn't been observed as of yet, only the extra
big compilation).
Fix it by deleting the relevant override_dependency call as part of
gstreamer_system_package() as well, which works as long as the syntax used for it
is standard in terms of whitespace usage.
For example gst-plugins-cdparanoia had 51 build units before due to this, but
only 2 after the fix.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>