gstreamer-meson.eclass: Avoid gl getting automatically enabled for base splits
This necessitates adding `gl` to GST_PLUGINS_NOAUTO in gst-plugins-base, or it gets
disabled there as well due to eclass additions being appended to the ebuild ones.
This is mainly for correctness, as the few base split builds didn't actually end up
building anything about GL extra due to this.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>