sci-biology/muscle: add 5.1.0
Major new release, major new build system that is majorly broken. <3
scientific code.

This package has not been updated since before the import of gentoo into
git, so I shall simply assume this major version is safe. From the
homepage, it is a "major rewrite", "based on new algorithms", which is
"much more accurate, is often faster, and scales to much larger datasets".

I think that's another way of saying "same old program but with fixes".

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>