games-puzzle/brainworkshop: add 5.0.3, enable py3.12
Note that this is 100% identical to the snapshot we we using for
5.0.2_p20230115 (not a single commit since), only difference is
that it's officially released and we can use the tag.

At same time, has no tests but py3.12 "seems" usable from quickly
trying runtime. I say usable given it's not perfect, it startups
more slowly and emits a warning, but beyond that it seems to run
normally so think it's fine to add it.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>