sci-biology/qrna: add 2.0.4
This is the 2006 edition of the software. It has a bunch of useful
fixes:
- it fixes Modern C porting issues
- it fixes LTO errors
- it incorporates the glibc-2.10 patch

just by upgrading! It's about time we upgrade, since it is 18 years
later.

Along the way we fix up the ebuild a bit:
- bump to EAPI 8
- acquire an actual SRC_URI, not just the gentoo mirror
- update the homepage along with the SRC_URI. The original homepage is
  dead, and the domain redirects to a new domain (whence comes the new
  distfile).

It has some shortcomings too:
- the distfile comes with free .svn directories, which have to be
  pruned.
- the distfile comes without the 80mb demos

To solve this, we use the old distfile (which I think only exists on
Gentoo's mirror network) and gate it behind a USE=examples flag, in case
anyone needs them.

Closes: https://bugs.gentoo.org/862285
Closes: https://bugs.gentoo.org/884283
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>