sci-libs/mumps: mark as LTO-unsafe, strict-aliasing unsafe
I would like to report this upstream. However, the only contact method
is their mailing list. And their mailing list may or may not work -- I
find it hard to tell because I've never seen the interface before, but
it is mostly in French and the bit that I recognized doesn't seem to
have working links to e.g. the archive.

Closes: https://bugs.gentoo.org/862903
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>