sci-electronics/geda: remove useless USE conditionals that never fired
For several reasons, this effort to control examples/docs via sed'ing
out Makefile.in did not work:
- the seds didn't apply
- the Makefile.in was ignored because it was regenerated via eautoreconf

... and also useless our outright wrong. The examples are 548kb, and the
docs are 156kb. Small files shouldn't be gated on USE flags.

But also, the docs USE flag was attempting to remove the part of the
build system that caused man pages to be shipped, in addition to the
HTML versions of the man pages. This must never be done, manpages are
too important. See QA Policy Guide 0305.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>