sci-libs/oc: mark as LTO-unsafe, strict-aliasing unsafe
Upstream is libdap. They exist. They have lots of maintained software,
it is simply that oc isn't one of them as far as I can tell.

What happened to it? I don't know. Maybe it's legacy abandoned software.
Their distfiles server still has it, but that's it.

Possibly we should delete the package instead...

Closes: https://bugs.gentoo.org/862906
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>