sys-apps/bat: fix build with >=libgit2-1.8.0
libgit2-sys crate has a upper bound for <1.8.0 and refuses
to build otherwise, but not seeing(?) any issues with it.

Just a quick emergency fix, will likely be fixed next libgit2-sys
version. Kept full version in the sed's path rather than a glob
to force re-checking this when the crate is bumped.

Alternatively could drop LIBGIT2_NO_VENDOR or set an upper bound
in RDEPEND, but does not seem necessary.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>