x11-wm/fvwm: remove pointless dependency on debianutils
This was originally added back in 2004, to support a patch called
fvwm-2.5.8-fvwm-bug-security.diff. It modified the fvwm-bug program with
the following diff as a backport from upstream:

```
-TEMP=/tmp/fvwm-bug.$$
+TEMP=`mktemp -q "/tmp/fvwm-bug.XXXXXX"`
```

Of course, in 2004 the mktemp program was not yet part of coreutils.
That came later on, in 2007. But back when this change was first
introduced, debianutils was the canonical source of this program.

17 years later, no one has ever noticed the reason for this dependency,
so here it sits, haunting us.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Matt Jolly <kangie@gentoo.org>