games-board/freedoko: add 0.8.0
Brings us up from 2017 to end of 2023. Drop hack for CXX std as it was
introduced in commit 91ecddb6c84583bc2839b5b510629880a763a045 for the
purpose of "Fix building with latest glibmm/libsigc++" and it seems
quite archaic as a temporary hotfix.

Reroll the patch, drop 2 data directories that were dropped upstream.

The new version handily fixes compiling with LTO.

Closes: https://bugs.gentoo.org/926051
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>