net-dns/djbdns: hide -Wincompatible-pointer-types problems
This is the worst solution to this problem, but our djbdns package isn't
djbdns at all, it's twenty patches in a trenchcoat. Maintenance for
things like this typically comes in the form of yet more downstream
patches, but invasive patches are essentially blocked by the fact that
they introduce conflicts with the others. There is at least one big
patch, the ipv6 patch, that is maintained by someone else, so it's not
as simple as basing all new patches on top of the old ones.

For lack of a better solution, we append -Wno-error for this warning.

Closes: https://bugs.gentoo.org/927539
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>