media-libs/gstreamer: drop unused USE=orc
I can't find any orc usage in the core module in recent history (including
autotools era 1.14), yet it even had pax-markings for it for tools, so I
assume it did use it long time ago?
It isn't used anymore and we actually get a QA warning about it after the
eclass fixups for false positives. While that QA warning still doesn't
cover it well for non-splits, it turns out it's actually right - it is not
used anywhere in the core module, so drop the USE flag and cleanup the
phases that aren't needed anymore after removing the orc conditional
pax-mark stuff.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>