profiles: mask media-libs/embree-4 for now
Did give it a quick try with minimal USE and (as noted in the bug)
embree3 and 4 do coexist without any changes and seems to me they
are indeed meant to be treated like a different SLOT or package.

	/usr/lib64/libembree3.so
	/usr/lib64/libembree4.so
	/usr/include/embree3
	/usr/include/embree4

(either way the soname changed, *if* not slotted it'd need a subslot)

Revdeps may need attention to depend on specific slot and select
properly if they can use either (assume blender does automagic?
currently does not depend on >=4 anyhow) so just doing a quick
mask until this is properly looked at.

Godot already depend explicitly on embree:3. If *really* had to
it could optionally switch to bundled embree rather than set
an upper bound if not slotted (or when :3 is to be removed).

Bug: https://bugs.gentoo.org/927833
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>