x11-drivers/nvidia-drivers: stabilize 550.67 for amd64
This is the 3rd iteration of the new 0/550 production branch
and it is about time to stabilize it. Note however that some
Lenovo laptop users may or may not experience random crashes
and need to stick to 0/535 for now (which is still supported
and not going anywhere for a while).

That aside, happy bday immolo :)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>