dev-games/godot: fixup shell completion files
Won't bother with godot*-runner.

Skipping revbump given this is a long build and most users are likely
not using shell commands and work through the GUI (which may explain
why the only report for this being broken was automated).

(missed QA myself given it only shows if bash-completion is installed)

Closes: https://bugs.gentoo.org/927855
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>