sys-devel/llvm: add workaround for ppc with < GCC 14
Pass -fno-ipa-sra -fno-ipa-modref -fno-ipa-icf on ppc with < GCC 14. We know
the problem is related to IPA and these 3 together help. Someone is welcome
to spend the time trying to narrow it down more, but applying this as a bandaid
which is better than nothing.

GCC 14 seems fine without this, but the "fix" bisects to something odd
just like the original "break" commit, so...

Bug: https://gcc.gnu.org/PR108250
Bug: https://bugs.gentoo.org/914888
Closes: https://bugs.gentoo.org/880677
Closes: https://bugs.gentoo.org/875776
Signed-off-by: Sam James <sam@gentoo.org>