sys-cluster/legion: remove conceptually invalid repository clone url
This clearly wasn't tested (or else it was a cunning form of troll) to
list the primary clone url as git://directoryname/reponame.git given
that the directory name "StanfordLegion" is not a valid hostname or IP
address.

If I had to hazard the wildest of wild guesses as to what this was
intended to do, it was intended to allow cloning from github.com over
your choice of git:// or https:// protocols. Which leaves open the
alternative question "why in G-ds name...".

In any case, it doesn't matter since github hasn't offered git://
support for quite some time, so even if it worked, it wouldn't work.
This change was made by github for the same reason that portage scolds
you for using it:

 * git-r3: git protocol is completely insecure and may render the ebuild
 * easily susceptible to MITM attacks (even if used only as fallback). Please
 * use https instead.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>