dev-qt/qtdeclarative: fix qtsvg automagic in 6.7+
Qt6Svg been checked by the build system for a long time but
was only used for unused examples and (manual) tests and so
was ignored.

...but 6.7 adds a svgtoqml tool which actually links with it
and is otherwise skipped without qtsvg.

Doesn't seem very interesting, so left off-by-default for now.
Could change if something needs it, albeit USE=svg is default
on desktop profiles either way.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>