app-arch/xz-utils: add/restore 5.4.2
This is the last release signed by Lasse Collin, the previous signer of xz-utils
releases.

Downgrade to this out of an abundance of caution. We are not aware of any issues
that *specifically* require this.

Note that the Manifest matches dfcc1f271fa3da8b8710c80737e85a7347f16ba0 from
when 5.4.2 was removed from ::gentoo in the past.

Bug: https://bugs.gentoo.org/928134
Signed-off-by: Sam James <sam@gentoo.org>