dev-qt/qt-creator: install Devel component with IUSE=plugin-dev
Tried to compile Qt Creator's HelloWorld plugin externally and it built
fine with the USE enabled.

Debated installing unconditionally, but it installs quite a lot of
extras. Unsure about the IUSE name choice but it should hopefully do
(did feel IUSE=devel would be a bit too generic sounding).

Went for a bit hacky solution (for now) given --component integrates
poorly with cmake.eclass and does not seem(?) other ebuilds have needed
to do this yet in ::gentoo for ideas.

Closes: https://bugs.gentoo.org/928423
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>