app-arch/xz-utils: depend on >=sec-keys/openpgp-keys-lassecollin-20230213 for verify-sig
Someone reported an expired key issue when verifying 5.4.2. That might be because
of an old version of gemato which didn't check whether the key was expired back
when the signature was made (since fixed) or because of an old version of
openpgp-keys-lassecollin.

Add a dep to sort out the latter given it's possible.

(Note that sec-keys/openpgp-keys-lassecollin hasn't changed in quite some time,
the latest version was added over a year ago in 81260457d76492e0e7a26241c6b594b4d35f869f.)

Signed-off-by: Sam James <sam@gentoo.org>